backend/requirements.txt
# GR PitIQ Backend Requirements
# Production dependencies for Flask API server

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
Werkzeug==3.0.1

# Database
supabase==2.3.0

# Environment & Configuration
python-dotenv==1.0.0

# ML/Data Science (for model serving)
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2

# Optional: Production Server (recommended)
gunicorn==21.2.0
gevent==23.9.1

# Monitoring & Logging
python-json-logger==2.0.7
